refactor(codex-comment-checker): split apply patch extraction
