+++ packages/omo-codex/plugin/components/comment-checker/src/core-values.ts
@@ -1,0 +1,11 @@
+export function getString(input: Record<string, unknown>, keys: string[]): string | undefined {
+	for (const key of keys) {
+		const value = input[key];
+		if (typeof value === "string") return value;
+	}
+	return undefined;
+}
+
+export function isRecord(value: unknown): value is Record<string, unknown> {
+	return typeof value === "object" && value !== null;
+}
